position model upd

+++ Models/Position.cs
@@ -5,6 +5,6 @@ namespace database_scaffold_asp_dot_net.Models
     {
         public int Id { get; set; }
         public string Title { get; set; }
-        public decimal Salary { get; set; }
+        public long Salary { get; set; }
     }
 }
